animated plots working
